refactor: move state into game_library
