+++ game/src/resources/mod.rs
@@ -8,7 +8,6 @@ mod cursor_position;
 mod fonts;
 mod plugin;
 mod return_to_state;
-mod state;
 
 pub mod buttons;
 pub mod movement;
@@ -16,4 +15,3 @@ pub mod style_prefab;
 
 pub use plugin::ElementalistResourcesPlugin;
 pub use return_to_state::ReturnToState;
-pub use state::*;
