+++ game/src/settings_menu/accessibility.rs
@@ -4,9 +4,10 @@ use bevy::prelude::*;
 use game_library::{
     font_resource::{ChangeFont, FontChoice, FontResource},
     settings::{next_font_family, AccessibilitySettings, SettingCategory, SettingChanged},
+    state::MenuState,
 };
 
-use crate::{despawn_with_tag, resources::style_prefab, resources::MenuState};
+use crate::{despawn_with_tag, resources::style_prefab};
 
 use super::{
     base::SettingsMenuEntity,
