+++ game/src/settings_menu/audio.rs
@@ -4,9 +4,10 @@ use bevy::prelude::*;
 use game_library::{
     font_resource::FontResource,
     settings::{SettingCategory, SettingChanged, VolumeSettings},
+    state::MenuState,
 };
 
-use crate::{despawn_with_tag, resources::style_prefab, resources::MenuState};
+use crate::{despawn_with_tag, resources::style_prefab};
 
 use super::{
     base::SettingsMenuEntity,
