+++ game/src/settings_menu/base.rs
@@ -3,7 +3,7 @@
 use bevy::{prelude::*, sprite::MaterialMesh2dBundle, window::PrimaryWindow};
 use game_library::colors::BACKGROUND_COLOR_50;
 
-use crate::resources::MenuState;
+use game_library::state::MenuState;
 
 /// An entity tag for ease of cleanup when the menu is disabled.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Component)]
