+++ game/src/splash_screen/plugin.rs
@@ -1,7 +1,8 @@
 use bevy::prelude::*;
 
 use super::{components::OnSplashScreen, scene::splash_setup, systems::countdown};
-use crate::{despawn_with_tag, resources::AppState};
+use crate::despawn_with_tag;
+use game_library::state::AppState;
 
 // This plugin will display a splash screen with Bevy logo for 1 second before switching to the menu
 pub struct SplashScreenPlugin;
