+++ game/src/splash_screen/systems.rs
@@ -1,7 +1,7 @@
 use bevy::prelude::*;
 
 use super::resources::SplashTimer;
-use crate::resources::AppState;
+use game_library::state::AppState;
 
 // Tick the timer, and change state when finished
 pub fn countdown(
